+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- ILoadable.IsLoading flag
+- ILoadable.CanLoad() function
+
+### Fixed
+- Loader component correctly loads only not loaded ILoadable instances
+
 ## [1.0.0] - 2025-09-04
 ### Added
 - Loader